+++ CHANGELOG.md
@@ -5,13 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-
+- added tbl_commands to system database
+- added command actions to myems-api
+- added command feature to myems-admin
 ### Changed
-
+- updated tranlations of myems-admin
+- changed theme of BarChart and MultiTrendChart in myems-web
+- 
 ### Fixed
 - fixed HTTPError Deprecated Warning in myems-api
 - fixed edit user issue in myems-api
-- 
+- fixed save button issue of energy item dialog in myems-admin
+
 ### Removed